+++ ubelt/util_path.py
@@ -590,3 +590,34 @@ class Path(_PathBase):
         """
         util_io.delete(self)
         return self
+
+    @classmethod
+    def ensure_appdir(cls, appname, *args, type='cache'):
+        """
+        Args:
+            appname (str): name with paths
+            *args[str] : optional subdirs
+            type (str): can be cache, config, or data
+
+        Example:
+            >>> import ubelt as ub
+            >>> print(ub.Path.ensure_appdir('ubelt', type='cache').shrinkuser())
+            >>> print(ub.Path.ensure_appdir('ubelt', type='config').shrinkuser())
+            >>> print(ub.Path.ensure_appdir('ubelt', type='data').shrinkuser())
+            ~/.cache/ubelt
+            ~/.config/ubelt
+            ~/.local/share/ubelt
+
+            >>> import pytest
+            >>> with pytest.raises(KeyError):
+            >>>     ub.Path.ensure_appdir('ubelt', type='other')
+        """
+        from ubelt import util_platform
+        if type == 'cache':
+            return cls(util_platform.ensure_app_cache_dir(appname, *args))
+        elif type == 'config':
+            return cls(util_platform.ensure_app_config_dir(appname, *args))
+        elif type == 'data':
+            return cls(util_platform.ensure_app_data_dir(appname, *args))
+        else:
+            raise KeyError(type)